Move components js and css files into folders
