+++ src/App.css
@@ -5,7 +5,7 @@ body {
   margin: 0;
 }
 
-.App {
+.app {
   display: flex;
   flex-direction: column;
   min-height: 100%;
